Move overflow rendering and state to OutputManager
